array-functions: The is2DArray predicate checks if (a) an input is an array and (b) the first item of that input is an array, too.
boolean-functions: The match function provides a more functional approach to case differentiation. The function accepts a variable number (or an array) of pairs, a default value and an input. Each pair consists of a predicate (or value) and a return value (or function to compute the return value from the input).